BUG: filter out NameErrors after unexpected errors
>>> res = func()   # func raises an unexpected ValueError
>>> res

The second line raises a NameError because res is not defined:
`func()` raised before the name `res` was bound.

So, suppress these name errors after an unexpected exceptions.
See the discussion in  https://github.com/scipy/scipy/pull/13116,
where this first came in. Also note that here we are attaching the
`had_unexpected_error` flag to the test, not the runner (which can
possibly be reused for other tests).