MAINT: friendlier error reporting from DebugDTRunner
Format the error to the output stream, then raise. This is useful
for e.g. fail-fast command line usage: details are printed out,
then the doctest.UnexpectedError propagates to the system and
stops the process. The latter error is not very useful for a user,
so print out the actual doctest failure manually.