ENH: filter PendingDeprecationWarnings in DTChecker want->a_want etc
When run with warnings->errors, these trigger a failure in e.g.
`scipy.sparse.linalg.lobpcg`, because:
- the abbreviation of `np.diag(np.arange(1000))` contains a `...`
- which is interpreted as an Ellipsis, hence a_want/a_got become
  object arrays + a PendingDeprecationWarning is emitted
- this warning turned to error triggers a test failure without
  that much of a useful message.

The error reporting might need to be improved indeed, but that's separate.
Currently, filtering out these errors is a correct thing to do
as long as DTChecker does `w_want = eval(want)` etc.