ENH: allow for local files
Doctests are run in a tempdir, so if a doctest tries to open
a local file, the file is not found---unless copied manually.

Thus add an entry to DTConfig, `config.local_resources`,
which is a dict where keys are `test.name`s and values
are lists of filenames to copy. The file names are relative
to `test.filename`, i.e. most likely the sourcefile/module being
doctested.